# WIP: Allow inlining method matches with unmatched type parameters